# API_Documentation.yaml
openapi: 3.1.0
info:
  title: Game Exchange API
  version: '1.0.0'
  description: |
    The Game Exchange API allows you to create an account and managed used retro video games

    ### Features of Game Exchange API:

    * Add games to your collection
    * Trade your games with others

servers:
  - url: http://localhost:8080
  - url: http://localhost:3000

tags:
  - name: User
  - name: Games
  - name: Offers

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

paths:
  /login:
    post:
      tags:
        - User
      summary: Login to an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string  
      responses:
        '200':
          description: Login Authorized
        '401':
          description: Unauthorized
        '400':
          description: Missing Fields
        '404':
          description: Account Absent
        '500':
          description: Server Error
  /register:
    post:
      tags:
        - User
      summary: Register an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: integer
                address:
                  type: string
      responses:
        '201':
          description: Account Created
        '400':
          description: Missing Fields
        '409':
          description: Already Exist
        '500':
          description: Server Error
  /update:
    post:
      tags:
        - User
      security:
        - bearerAuth: []
      summary: Update user information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                password:
                  type: integer
                address:
                  type: string
      responses:
        '204':
          description: Updated
        '400':
          description: Mongo Error
        '500':
          description: Server Error
  /games:
    get:
      tags:
        - Games
      security:
        - bearerAuth: []
      summary: View games
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '200':
          description: Fetched
        '500':
          description: Server Error
    post:
      tags:
        - Games
      security:
        - bearerAuth: []
      summary: Add a new game to the exchange list
      requestBody: 
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                publisher:
                  type: string
                yearPublished:
                  type: integer
                system:
                  type: string
                condition:
                  type: string
      responses:
        '201':
          description: Game added
        '400':
          description:  Incorrect body
  /games/{gameId}:
    put:
      tags:
        - Games
      security:
        - bearerAuth: []
      summary: Replaces all properties of a game
      parameters:
      - name: gameId
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                publisher:
                  type: string
                yearPublished:
                  type: integer
                system:
                  type: string
                condition:
                  type: string
      responses:
        '202':
          description: Game Edited
        '404':
          description: Not Found
        '400':
          description: Incorrect Body
        '500':
          description: Server Error
    patch:
      tags:
        - Games
      security:
        - bearerAuth: []
      summary: Replaces certain properties of a game
      parameters:
      - name: gameId
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                publisher:
                  type: string
                yearPublished:
                  type: integer
                system:
                  type: string
                condition:
                  type: string
      responses:
        '202':
          description: Game Edited
        '404':
          description: Not Found
        '400':
          description: Incorrect Body
        '500':
          description: Server Error
    delete:
      tags:
        - Games
      security:
        - bearerAuth: []
      summary: Remove a game from your exchange list
      parameters:
      - name: gameId
        in: path
        required: true
        schema:
          type: integer
      responses:
        '202':
          description: Game removed from exchange list
        '404':
          description: Game not found
  /games/offer:
    get:
      tags:
        - Offers
      security:
        - bearerAuth: []
      summary: View all trade offers
      responses:
        '200':
          description: Request Complete
        '500':
          description: Server Error
    post:
      tags:
        - Offers
      security:
        - bearerAuth: []
      summary: Send trade offer
      in: path
      required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                offeredId:
                  type: integer
                desiredId:
                  type: integer
      responses:
        '200':
          description: Offer Sent
        '400':
          description: Invalid Inputs
        '500':
          description:  Server Eror
  /games/offer/{offerId}:
    post:
      tags:
        - Offers
      security:
        - bearerAuth: []
      summary: Accept an trade offer
      parameters:
      - name: offerId
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Trade Complete
        '400':
          description: Trade Error
        '500':
          description: Server Error